created server side file for admin
